Move RAWBLOCK functions out of overlay3.c into ml3think.c
Per review: the zako AI/think state machine (0x800D2C20-0x800D3B5C, with the
367C/3848/3ACC switch tables and its rodata strings) now lives in its own
ml3think.c; overlay3.c is data-only (.data + .bss). Byte-neutral split:
ml3think.obj provides .text+.rodata, overlay3.obj provides .data+.bss.
